puts all Postgres related code in a single file
