+++ dashboard/src/t5gweb/create_app.py
@@ -6,7 +6,7 @@ from flask import Flask
 from prometheus_flask_exporter import PrometheusMetrics
 
 from . import api, t5gweb, ui
-from .db import create_tables
+from .postgres_db import create_tables
 
 
 def create_app(test_config=None):
