--- dashboard/src/t5gweb/db.py
@@ -1,16 +1,0 @@
-# db.py - Database setup
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker, DeclarativeBase
-import os
-
-DATABASE_URL = os.getenv("DATABASE_URL", "postgresql://postgres:secret@postgresql/dashboard")
-
-engine = create_engine(DATABASE_URL)
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-class Base(DeclarativeBase):
-    pass
-
-# Create tables - this will be called after models are imported
-def create_tables():
-    Base.metadata.create_all(bind=engine)
