--- dashboard/src/t5gweb/dependencies.py
@@ -1,9 +1,0 @@
-from t5gweb.db import SessionLocal  # Import SessionLocal
-# dependencies.py - Dependency injection
-
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
